+++ ProcessMemory.Library/Properties/AssemblyInfo.cs
