+++ ProcessMemory.sln
@@ -1,19 +1,36 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.136
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28729.10
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProcessMemory", "ProcessMemory\ProcessMemory.csproj", "{35B5C872-D829-428F-9DFA-8EF2B4DE194C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProcessMemory.Library", "ProcessMemory\ProcessMemory.Library.csproj", "{35B5C872-D829-428F-9DFA-8EF2B4DE194C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProcessMemory.Test", "ProcessMemory.Test\ProcessMemory.Test.csproj", "{ED9F5080-993C-4BA4-9078-47A5153449CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
+		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Debug|x64.ActiveCfg = Debug|x64
-		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Release|x64.ActiveCfg = Release|x64
-		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Release|x64.Build.0 = Release|x64
+		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Debug|x64.Build.0 = Debug|Any CPU
+		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Release|x64.ActiveCfg = Release|Any CPU
+		{35B5C872-D829-428F-9DFA-8EF2B4DE194C}.Release|x64.Build.0 = Release|Any CPU
+		{ED9F5080-993C-4BA4-9078-47A5153449CB}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{ED9F5080-993C-4BA4-9078-47A5153449CB}.Debug|Any CPU.Build.0 = Debug|x64
+		{ED9F5080-993C-4BA4-9078-47A5153449CB}.Debug|x64.ActiveCfg = Debug|x64
+		{ED9F5080-993C-4BA4-9078-47A5153449CB}.Debug|x64.Build.0 = Debug|x64
+		{ED9F5080-993C-4BA4-9078-47A5153449CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ED9F5080-993C-4BA4-9078-47A5153449CB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ED9F5080-993C-4BA4-9078-47A5153449CB}.Release|x64.ActiveCfg = Release|x64
+		{ED9F5080-993C-4BA4-9078-47A5153449CB}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
